requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
ale-py==0.7.5
async-timeout==4.0.2
attrs==21.4.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.28
docopt==0.6.2
easybuild-easyblocks==4.4.1
easybuild-easyconfigs==4.4.1
easybuild-framework==4.4.1
fasteners==0.17.3
fonttools==4.33.0
frozenlist==1.3.0
glfw==2.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gym==0.21.0
-e git+https://github.com/shufflebyte/gym-nav2d@445a28b53eccf46ff0519ce8afe807dcca20f13a#egg=gym_nav2d
gym-notices==0.0.6
idna==3.3
imageio==2.19.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
importlib-resources==5.7.1
joblib==1.2.0
Js2Py==0.71
kiwisolver==1.4.2
labmaze==1.0.5
lockfile==0.12.2
lxml==4.8.0
lz4==4.0.1
Markdown==3.3.6
matplotlib==3.5.3
mujoco==2.2.0
mujoco-py==2.1.2.14
multidict==6.0.2
multipledispatch==0.6.0
networkx==2.8.3
nose==1.3.7
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
patchelf==0.14.5.0
Pillow==9.1.0
pipwin==0.5.2
protobuf==3.20.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.21
pygame==2.1.2
pyglet==1.5.26
pyjsparser==2.7.1
PyOpenGL==3.1.6
pyparsing==2.4.7
PyPrind==2.11.3
pyrr==0.10.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
recordtype==1.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sb3-contrib==1.5.0
scikit-image==0.19.2
scikit-learn==1.1.3
scipy==1.8.1
six==1.16.0
soupsieve==2.3.2.post1
stable-baselines3==1.6.2
swigibpy==0.4.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.11.0
torch-tb-profiler==0.4.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.1
yarl==1.7.2
zipp==3.8.0
